pyproject.toml
[project]
name = "kingfisher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai>=1.70.0",
    "python-dotenv>=1.2.2",
]
